Basic Attack
Shooting projectiles have been added, you can use IJKL keys to shoot out projectiles(if anyone has a controller please test if this work with a controller as well thanks). Each button shoot out a projectile with a string value to it <- Has not tested, need enemies with corresponding shield.